work: task persistence tracking
As scaffolding, +merge-diff deleted all task history for a given file in state, but to retain metadata for a task that persists across file updates, +merge-diff needs to check each task to know whether it's been retained or not.

To avoid complications, +merge-diff simply compares tasks cords from the previous file to the update, and if it finds a match, it retains it as is, including its old task-id. This is so any meta data (e.g. contributor assignment, %fund link, etc.) connected to it is retained.

While this feels a bit fragile - if the maintainer merely changes the wording of an existing task, %work will recognize it as a new one, other methods such as line tracking is more fragile, and would require far more complex tracking functions with in %work.